refactor(AttendanceInit): InitService 삭제 및 Init 도메인 생성
- Service 레이어 보다는 객체 사이에 역할에 집중할 수 있도록 변경했습니다.
